+++ app/src/main/java/com/jerboa/ui/components/common/Navigation.kt
@@ -151,4 +151,4 @@ fun NavController.toLookAndFeel() = navigate(Route.LOOK_AND_FEEL)
 
 fun NavController.toAbout() = navigate(Route.ABOUT)
 
-fun NavController.toCrashes() = navigate(Route.CRASHES)
+fun NavController.toCrashLogs() = navigate(Route.CRASH_LOGS)
