+++ app/src/main/java/com/jerboa/ui/components/settings/about/AboutActivity.kt
@@ -32,7 +32,7 @@ import com.alorma.compose.settings.ui.SettingsMenuLink
 import com.jerboa.R
 import com.jerboa.openLink
 import com.jerboa.ui.components.common.SimpleTopAppBar
-import com.jerboa.ui.components.common.toCrashes
+import com.jerboa.ui.components.common.toCrashLogs
 
 const val githubUrl = "https://github.com/dessalines/jerboa"
 const val jerboaMatrixChat = "https://matrix.to/#/#jerboa-dev:matrix.org"
@@ -99,14 +99,14 @@ fun AboutActivity(
                     },
                 )
                 SettingsMenuLink(
-                    title = { Text(stringResource(R.string.crashes)) },
+                    title = { Text(stringResource(R.string.crash_logs)) },
                     icon = {
                         Icon(
                             imageVector = Icons.Outlined.Build,
                             contentDescription = null,
                         )
                     },
-                    onClick = { navController.toCrashes() },
+                    onClick = { navController.toCrashLogs() },
                 )
                 SettingsMenuLink(
                     title = { Text(stringResource(R.string.settings_about_developer_matrix_chatroom)) },
